fix openvpn connection

+++ world/bkeka/vpn_tools.py
@@ -19,9 +19,8 @@ if platform.system().lower().startswith('win'):
     PIPE_NOWAIT = wintypes.DWORD(0x00000001)
 elif platform.system().lower().startswith('lin') or platform.system().lower().startswith('dar'):
     from fcntl import fcntl, F_GETFL, F_SETFL
-    
 
-SUCCESS_MESSAGE="Initialization Sequence Completed"
+SUCCESS_MESSAGE = "Initialization Sequence Completed"
 OPENVPN_COMMAND = "openvpn"
 VPN_SUCCESS = 1
 VPN_ERROR = 0
@@ -30,9 +29,11 @@ MAX_RETRIES = 5
 # Global variable for VPN subprocess
 openvpn_process = None
 
+
 class OpenVPNProcessExceptions(Exception):
-   """Raised for different internal errors"""
-   pass
+    """Raised for different internal errors"""
+    pass
+
 
 def pipe_no_wait(process):
     """ pipefd is a integer as returned by os.pipe """
@@ -51,17 +52,18 @@ def pipe_no_wait(process):
         return True
     elif platform.system().lower().startswith('lin') or platform.system().lower().startswith('dar'):
         # set the O_NONBLOCK flag of openvpn_process.stdout file descriptor:
-        flags = fcntl(process.stdout, F_GETFL) # get current p.stdout flags
+        flags = fcntl(process.stdout, F_GETFL)  # get current p.stdout flags
         fcntl(process.stdout, F_SETFL, flags | O_NONBLOCK)
         return True
+
 
 def start_openvpn_subprocess(config_file, credentials_file):
     # Connect to vpn
     print("----->", config_file)
-    process = Popen(['sudo', OPENVPN_COMMAND, config_file], stdout=PIPE, stderr = PIPE, shell = False)
+    process = Popen(['sudo', OPENVPN_COMMAND, config_file], stdout=PIPE, stderr=PIPE, shell=False)
     if not pipe_no_wait(process):
         raise OpenVPNProcessExceptions("Failed to set file descriptor non blocking")
-    
+
     return process
 
 
@@ -72,23 +74,20 @@ def get_config_file(vpn_config_dir):
             return x
     return None
 
+
 def close_vpn_connection():
     global openvpn_process
-    # check if child process has terminated
-    if openvpn_process is not None:
-        returncode = openvpn_process.poll()
-        if returncode is not None:
-            openvpn_process.kill()
-            openvpn_process = None
-            sleep(3)
-    # zombie openvpn process and open tun devices in the wild
-    process = Popen(['sudo', 'killall', OPENVPN_COMMAND], stdout=PIPE, stderr=PIPE, shell=False)
-    if not pipe_no_wait(process):
-        raise OpenVPNProcessExceptions("Failed to set file descriptor non blocking")
+    if openvpn_process is not None and openvpn_process.poll() is not None:
+        openvpn_process.kill()
+        openvpn_process = None
+    # 'sudo killall openvpn' does something nasty to VPN provider connection
+    # process = Popen(['sudo', 'killall', OPENVPN_COMMAND], stdout=PIPE, stderr=PIPE, shell=False)
+    # if not pipe_no_wait(process):
+    #    raise OpenVPNProcessExceptions("Failed to set file descriptor non blocking")
 
 
 def connect_to_vpn(vpn_config_dir, vpn_credentials_file):
-    logger = logging.get_logger(__name__+'.log', __name__+'.log')
+    logger = logging.get_logger(__name__ + '.log', __name__ + '.log')
     vpn_config_file = None
     init_success = False
     retry_no = 0
